+++ packages/common/src/content/hostedServiceUtils.ts
@@ -1,7 +1,7 @@
 import { IFeatureServiceDefinition } from "@esri/arcgis-rest-feature-layer";
 import { IItem } from "@esri/arcgis-rest-portal";
 import { IHubEditableContent } from "../core/types/IHubEditableContent";
-import { isHostedAgolService } from "./isHostedAgolService";
+import { isAGOHostedService } from "./isAGOHostedService";
 
 /**
  * DEPRECATED: This will be removed in the next breaking version. Use "isHostedFeatureServiceMainItem" instead.
@@ -90,7 +90,7 @@ export function isAGOFeatureServiceUrl(url: string): boolean {
   // NOTE: if this is the URL of a proxied service it will fail the "is hosted" check
   // even if the proxy points to a hosted service (which should be rare)
   return (
-    !!url && isHostedAgolService(url) && FEATURE_SERVICE_URL_REGEX.test(url)
+    !!url && isAGOHostedService(url) && FEATURE_SERVICE_URL_REGEX.test(url)
   );
 }
